matrix-appservice-irc: Refactor + hardening
The pre-start section is factored out to its own Systemd service. The hardening options
are moved accordingly to avoid the introduction of redundancy between both services. Some
hardening settings also got tweaked but they are all disabled at the moment. Getting them
back to work (a few additions might have been over-eager) is left as an exercise to the reader.